cloud_service/build_context/requirements.txt

flask
ultralytics
pillow
gunicorn
firebase-admin
google-cloud-firestore
google-cloud-storage